git-svn-id: https://gnunet.org/svn/GNUnet@8617 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/applications/dhtlog_mysql/dhtlog_mysql.c
@@ -33,7 +33,7 @@
 
 #define DEBUG_DHTLOG GNUNET_YES
 
-static unsigned long long max_varchar_len;
+static unsigned long max_varchar_len;
 
 static char *blank;
